fix(ci): python-coverage-shards line-limit failure
Split the oversized PR monitor coverage-edge test file so the maintainability guardrail no longer reports part 008 above 1500 lines.
